#102 [BUGFIX] 메인페이지 한글깨짐 오류 수정
